Added $message->react_delete()
There are three types, (first box). The first type is 'all', which doesn't require any params after it and will delete ALL reactions attached to the message. Second is 'me', which requires the second param, $emoticon, to function. it removes the reaction that the bot has placed., and the third is 'id', which requires a userid to function and $emoticon to remove.