Added search ability in $guild->search($options);
Allows you to search for certain things in a guild's past messages.

$options = [
'has' => [],
'content' => '',
'author_id' => [],
'channel_id' => [],
'mentions' => [],
'max_id' => '',
'min_id' => ''
];

Mentions AND author_id can be a `Discord\Parts\User\User Object`, userid, or an array mixed with both.

Channel can be a 'Discord\Parts\Channel\Channel Object`, channelid, or an array mixed with both.

Content is a string that can be anything.

Has is an array of different things, `['link', 'embed', 'file', 'video', 'image', 'sound']` to be specific. So it would be `$options = ['has' => ['link', 'embed']];`

max_id and min_id is a snowflake of a messageid/date or something, idk.

For the search "During", you would just make the max_id the end of a day, and min_id the start of the day. For example:

2017-02-08 would be `min_id = 278751766118400000` and `max_id = 279114153984000000`