Added customizable custom_id feature (#583)
* added customizeable custom_id feature

Changed the constructor to expect a nullable param $customId. If the param is given, the custom_id is set to the given $customId. Otherwise its set with $this->generateUuid()
Added getter and setter for the custom_id to read and write them with the object instance.

* Bugfix SelectMennu.php

Removed the duplicated new() Method