Declare and Implement getFillableAttributes()
There wasn't a way to get a part's $fillable array from outside of the class, which can make it tricky to dynamically create an object in certain use-cases. This resolve it by allowing the fillable property to be retrieved with a function. The main downside to this method is that the function has to be redeclared in every part that has a non-empty fillable array because of the usage of reflection, but I think that's a small price to pay for future maintainability.