DiscordCommandClient.php: don't add trailing space to usage
The trailing space added to the `usage` property of a command does not
serve any purpose. It is a holdover from when `usage` was being used
to construct a full command-line string:

    $helpString = "{$prefix}{$this->command} {$this->usage}- {$this->description}\r\n";

This was removed in b3e0e900ce in favor of printing `usage` as-is, and
so the added space no longer makes any sense.