Fix callable typehints
The `callable` keyword does not refer to a class or global function, so
it should not be prefixed with a backslash. We replace `\Callable` with
`callable` (the capitalization change is cosmetic, since it is standard
for non-class types to be lowercase).

This was causing errors in static analysis for dependent libraries when
calling into the methods that used the `\Callable` typehint.