The actual pause/unpause bug fix
It seems that $delay was negative after unpausing because $loops wasn't getting incremented when the stream was paused. This unexpected side effect seemed to be cause the issue. My """fix""" bypassed the $loop which is why it seemed to of worked but I neglected to test on windows (woops). This fix however was tested on Windows and Linux. I also had the person who said my original """fix""" broke audio streams test and they said it works.