ExCollectionInterface (#1305)
This includes the newly introduced functions that were added as part of recent enhancements. We do not use these functions internally within the library, so rather than update the original interface to include them as requirements, it makes sense for end-users to utilize the ExCollectionInterface instead to take advantage of the full functionality.