Automatically wrap SelectMenus in ActionRow
The old behavior did this directly in SelectMenu::jsonSerialize, such that anyone adding a SelectMenu manually to an ActionRow (as is the correct usage) would end up with an ActionRow nested inside of another ActionRow, which would always result in Discord rejecting with a `50035 Invalid Form Body` error message.

The new behavior moves this logic directly to Container::addComponent and MessageBuilder::addComponent by creating a new ActionRow and adding the SelectMenu to it. preserving the original usage without breaking containers.