Fix broken caching for interactions
`isset($interaction->attributes['member'])` would have worked as expected because it checks the array directly, but `isset($interaction->member)` does not work the same way, since it relies on a magic method (`__get`) that only gets invoked if the property is accessed, not when `isset()` is used.
This means the function is effectively misimplemented, because `__get` cannot influence `isset()` checks.